Audio: Use AP-side CPort ID value for mgmt_cport
The gb_connection structure contains the CPort ID values
for the AP side and for the module side.  The AP side
CPort ID value is in 'hd_cport_id' and the module side
module side CPort ID value is in 'intf_cport_id'.

Use the AP side value for in the Greybus audio driver.

Signed-off-by: Mark Greer <mgreer@animalcreek.com>